# Optimizing BigUint and Bigint multiplication with the Toom-3 algorithm