@kompass 2017-08-14
Hum I used a feature not available in the 1.8 again.