@kompass 2017-08-15
After these two optimizations :

```
running 4 tests
test multiply_0        ... bench:         103 ns/iter (+/- 4)
test multiply_1        ... bench:      15,405 ns/iter (+/- 1,377)
test multiply_2        ... bench:   1,029,812 ns/iter (+/- 29,265)
test multiply_3        ... bench:   2,634,657 ns/iter (+/- 37,420)
```
I optimized `mac_digit` too because Valgrind measured that mac3 spent 28% of its time calling mac_digit.