@cuviper 2017-09-02
Ping?

I'm not looking for extensive documentation, just some brief comments to explain what this code is doing.