@cuviper 2017-09-20
Well, I added some brief comments and made a few more performance tweaks of my own.  Let's test it!

bors r+